requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3
django-crontab==0.7.1
djangorestframework==3.14.0
djongo==1.3.3
dnspython==2.3.0
idna==3.4
pymongo==3.12.3
pytz==2023.3
requests==2.28.2
sqlparse==0.2.4
tzdata==2023.3
urllib3==1.26.15
waitress==2.1.2
